dynamic proxy instance spacing and minor refactoring
